test: Split encode parts out to make them usable from unit and int tests
`test_types` is structurally an integration test. But effectively does
not test the integration of services/components. Instead its tests are
actually unit tests for `types.rs`. So the aim is to turn `test_types`
into unit tests. But this is not possible, as they rely on
`encode_values` which lives in `tests/support` and hence not accessible
from unit tests.

But we cannot pull these `encode` functions into `types.rs`' unit tests,
as they are also used parser integration tests and `bench_basic`.

In order to allow accessing the `encode` functions from both unit tests,
integration tests, and benches, we move them into a `shared` module,
which we can import from all these places.

This will allow to turn `test_types` into unit tests in follow-up
commits.

This is part of #2249
