+++ redis/tests/support/shared/mod.rs
@@ -1,0 +1,3 @@
+//! Tooling for tests that should be usable from both unit and integration tests
+
+pub mod encode;
